trader/requirements.txt
tradingview-ta
python-telegram-bot
emoji
psycopg2-binary==2.8.6
flask==1.1.2